requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-multipart==0.0.20
requests==2.32.3
Pillow==10.4.0
numpy==1.26.4
rembg[gpu]==2.0.59
onnxruntime==1.19.2
